Revert "UPSTREAM: driver core: Don't do deferred probe in parallel with kernel_init thread"
This reverts commit 301a95087dc81a9d389e300cef4b00834091387b.

Conflicts:
	drivers/base/base.h

To fix the non-bootable state of kernel when building drivers in instead of modules.
The order of loading socinfo and smem using drivers fail because the smem driver is
being probed too late in the process. For now easy solution is to revert this commit.
Some kind of deperndency based order would be best.